+++ packages/config/package.json
@@ -52,14 +52,17 @@
     "fs-extra": "^7.0.1",
     "invariant": "^2.2.4",
     "jest-message-util": "^25.1.0",
+    "plist": "^3.0.1",
     "resolve-from": "^5.0.0",
     "slugify": "^1.3.4",
+    "xcode": "^2.1.0",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
     "@expo/babel-preset-cli": "0.2.6",
     "@types/invariant": "^2.2.30",
     "@types/xml2js": "^0.4.5",
+    "@types/plist": "^3.0.1",
     "memfs": "^2.15.5"
   },
   "publishConfig": {
